src/requirements.txt
fastapi
uvicorn
kafka-python
faker
numpy